Add a new cog: heropoll.
Example:

A new poll is created:
  heropoll Lina, Zeus

The bot prints a message, for example:
  Which hero should I play?

  1.  Lina
  2.  Zeus

  Type the number to vote!

Then the users respond by typing a number in the channel.
If the bot has "Manage Messages" permissions, it will delete users votes after
counting them to avoid clutter.
Regardless of whether the vote message was deleted, the bot will whisper the
user so that they know their vote was counted (perhaps this should be changed?).

Finally, the poll author ends the poll:
  heropoll stop
  heropoll abort (same as stop, but without printing results)

And the bot prints the results:
  Which hero should I play?

  1.  Lina (3 votes)
  2.  Zeus (7 votes)

  The winner is Zeus with 7 votes!

Limitations:
 - Due to users typing responses, only 1 poll per channel is supported at a
   time. Using something like reactions would remove this limitation.
 - The question cannot yet be customized; it's always "What hero should I play?"
 - No attempt has been made to restrict the creation of polls to specific roles,
   so anyone can create a hero poll. Since there's only 1 poll allowed per
   channel and only the author can stop a poll, any user can effectively DoS the
   channel. This should be easy to fix, but this change already has way too much
   code in it.